pyproject.toml
[project]
name = "o11y-tools"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "python-dotenv>=1.0",
]

[project.scripts]
o11y-up = "o11y_tools.cli:main"

[tool.uv]
# optional: keep uv venv in-repo
# venv is usually .venv automatically
